refactor(steam): colocate app id in tamagotchi and server
