+++ apps/server/package.json
@@ -48,7 +48,6 @@
     "@proj-airi/drizzle-orm-browser-migrator": "catalog:",
     "@proj-airi/server-schema": "workspace:*",
     "@proj-airi/server-sdk-shared": "workspace:*",
-    "@proj-airi/stage-shared": "workspace:^",
     "better-auth": "catalog:",
     "cac": "catalog:",
     "drizzle-orm": "catalog:",
