+++ apps/stage-tamagotchi/src/main/services/steam/app-id.ts
